Registration Update + DB Clean
During this change, the registration form was created as fully functional and can now reach out to our registration API running via DotNet to create new users in the database. The database was cleaned out to clear out any bad data (such as a username previously created with a blank string).