Routing, Toasting, Additional Components
Additional components for the UI were started. Angular routing was implemented to better handle when a user clicks a link to load a different form on the SPA. Toasting was implemented to handle error messages in a more graceful way (as opposed to logging the error to the console for any user to see).