Implemented Automapper to map new DTO's
New DTO's were created to ensure each user was tied to multiple images, and that the proper values were returned from the API on certain requests (excluding things such as the PasswordHash and PasswordSalt used). AutoMapper was utilized as a helper to map these DTO's to Entities and return relevant information with ease.