Set up pagination in API and Client
Pagination was added in the API and the client to help reduce the workload of the server should there be a large user set. Instead, pagination was added to load a maximum of 50 users at a time and only display those 50 to reduce workload and also display a reasonable number of users.