Added additional sorting functionality
Sorting by last active and most recently created users is now possible in the application. Changes were made to include an additional claim in the JWT to give a user's ID directly back to the server via the token, user's most recent activity is now updated on every request to the server, and the client can now allow for switching between the two filters.